constraints take array instead of vector as input type
